Removed as many warnings as possible

+++ entitylist.cpp
@@ -66,7 +66,7 @@ void EntityList::clear()
 
 void EntityList::renderAll(Gfx &gfx)
 {
-	for (int i = 0; i < entities.size(); i++)
+	for (size_t i = 0; i < entities.size(); i++)
 	{
 		entities.at(i)->render(gfx);
 	}
